requirements.txt
flatland-rl
networkx
ipykernel
jupyter
nb_conda_kernels
pytest
pytorch
torchvision
scipy
black
isort
pre-commit